Adds handling of matches for WebView.
